Changing table status after bill payment is done.
Changing discount to amount instead of percentage.

Related work items: #793

+++ src/Libraries/SmartRestaurant.Application/Common/Dtos/BillDtos/BillDishDto.cs
@@ -7,7 +7,7 @@
         public string Name { get; set; }
         public float InitialPrice { get; set; }
         public float UnitPrice { get; set; }
-        public int Discount { get; set; }
+        public float Discount { get; set; }
         public float Quantity { get; set; }          
     }
 }
